add a module for system commands
